app/requirements.txt
fastapi
sqlalchemy
python-dotenv
uvicorn